SFT 冒烟测试通过 + 环境兼容修复 + 窗口式验证
环境/兼容 (本机 2×A100 80G, 冒烟测试实测):
- 新环境 mer2026: torch 2.1.2+cu121 / torchvision 0.16.2 / torchaudio 2.1.2 /
  transformers 4.40.2 / peft 0.11.1 / setuptools<81 / pytorchvideo 兼容
- torch 版本检查 startswith('2.1.0')→'2.1' (3 文件)
- ImageBind/data.py: pytorchvideo 容错导入 + 音频用自实现均匀切片
- dataloader_utils: PrefetchLoader/MultiIterLoader 补 __next__/__iter__/reset (val 每轮重跑)
- affectgpt.py: LLM 加载改 fp32 (fp16+scaler 报 unscale 错, 纯 fp16 训练 NaN)

验证逻辑 (窗口式):
- 每轮: 轻量 val CE loss (eval_val_loss, 记录到 log.txt + _val_loss_history)
- 每 eval_interval 轮: 对窗口内 val loss 最低的 epoch 加载其 checkpoint 做生成式 EW-F1,
  验证后恢复当前权重, 再按 EW-F1 选最优存 checkpoint_best
- task.before_evaluation 解包 reorg 后的 list
- pipeline_io.md: 补 Prompt 设计问答 + Q-Former 特征流
- setup_gpu_env.sh: 更新为已验证版本 + 冒烟测试步骤
- 新增 sft_smoke_test.py (12 训练 + 6 验证, 2 epoch, 全链路冒烟)

+++ my_affectgpt/models/affectgpt.py
@@ -86,13 +86,13 @@ class AffectGPT(Blip2Base):
         if llama_model_name in ['Baichuan2']:
             self.llama_model = AutoModelForCausalLM.from_pretrained(
                 config.PATH_TO_LLM[llama_model_name],
-                torch_dtype=torch.float16,
+                torch_dtype=torch.float32,
                 trust_remote_code=True
             )
         else:
             self.llama_model = AutoModelForCausalLM.from_pretrained(
                 config.PATH_TO_LLM[llama_model_name],
-                torch_dtype=torch.float16
+                torch_dtype=torch.float32
             )
 
         for name, param in self.llama_model.named_parameters():
